requirements.txt
torchmetrics==1.8.2
numpy==2.3.4
pandas==2.3.3
pillow==11.3.0
tqdm==4.67.1
PyYAML==6.0.3